backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
torch==2.9.1
python-dotenv==1.0.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.0

# Optional: Supabase client (for database integration later)
supabase==2.10.0
